# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Business OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ClientStatus = (typeof ClientStatus)[keyof typeof ClientStatus];

export const ClientStatus = {
  active: "active",
  inactive: "inactive",
  prospect: "prospect",
} as const;

export interface Client {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  company: string | null;
  status: ClientStatus;
  createdAt: string;
  updatedAt: string;
}

export type CreateClientBodyStatus =
  (typeof CreateClientBodyStatus)[keyof typeof CreateClientBodyStatus];

export const CreateClientBodyStatus = {
  active: "active",
  inactive: "inactive",
  prospect: "prospect",
} as const;

export interface CreateClientBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  status?: CreateClientBodyStatus;
}

export type UpdateClientBodyStatus =
  (typeof UpdateClientBodyStatus)[keyof typeof UpdateClientBodyStatus];

export const UpdateClientBodyStatus = {
  active: "active",
  inactive: "inactive",
  prospect: "prospect",
} as const;

export interface UpdateClientBody {
  name?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  status?: UpdateClientBodyStatus;
}

export type DealStatus = (typeof DealStatus)[keyof typeof DealStatus];

export const DealStatus = {
  new: "new",
  negotiation: "negotiation",
  won: "won",
  lost: "lost",
  on_hold: "on_hold",
} as const;

export interface Deal {
  id: number;
  title: string;
  value: number;
  status: DealStatus;
  /** @nullable */
  clientId: number | null;
  /** @nullable */
  clientName: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateDealBodyStatus =
  (typeof CreateDealBodyStatus)[keyof typeof CreateDealBodyStatus];

export const CreateDealBodyStatus = {
  new: "new",
  negotiation: "negotiation",
  won: "won",
  lost: "lost",
  on_hold: "on_hold",
} as const;

export interface CreateDealBody {
  title: string;
  value: number;
  status?: CreateDealBodyStatus;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type UpdateDealBodyStatus =
  (typeof UpdateDealBodyStatus)[keyof typeof UpdateDealBodyStatus];

export const UpdateDealBodyStatus = {
  new: "new",
  negotiation: "negotiation",
  won: "won",
  lost: "lost",
  on_hold: "on_hold",
} as const;

export interface UpdateDealBody {
  title?: string;
  value?: number;
  status?: UpdateDealBodyStatus;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  deadline: string | null;
  /** @nullable */
  dealId: number | null;
  /** @nullable */
  clientId: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyStatus =
  (typeof CreateTaskBodyStatus)[keyof typeof CreateTaskBodyStatus];

export const CreateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  status?: CreateTaskBodyStatus;
  priority?: CreateTaskBodyPriority;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  dealId?: number | null;
  /** @nullable */
  clientId?: number | null;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  dealId?: number | null;
  /** @nullable */
  clientId?: number | null;
}

export interface DashboardSummary {
  totalClients: number;
  totalDeals: number;
  totalRevenue: number;
  wonDeals: number;
  openTasks: number;
  activeClients: number;
}

export interface DealPipelineItem {
  status: string;
  count: number;
  value: number;
}

export interface ActivityItem {
  id: string;
  type: string;
  description: string;
  createdAt: string;
}

export interface SuggestReplyBody {
  clientMessage: string;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  context?: string | null;
}

export interface SuggestReplyResponse {
  reply: string;
}

export interface AnalyzeDealBody {
  /** @nullable */
  dealId?: number | null;
  title: string;
  value: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  clientName?: string | null;
}

export interface AnalyzeDealResponse {
  probability: number;
  reasoning: string;
  recommendations: string[];
}