# lib/api-zod/src/generated/types/analyzeDealBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Business OS API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeDealBody {
  /** @nullable */
  dealId?: number | null;
  title: string;
  value: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  clientName?: string | null;
}