# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Business OS API
 * OpenAPI spec version: 0.1.0
 */
import type { DealStatus } from "./dealStatus";

export interface Deal {
  id: number;
  title: string;
  value: number;
  status: DealStatus;
  /** @nullable */
  clientId: number | null;
  /** @nullable */
  clientName: string | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}