# lib/api-zod/src/generated/types/suggestReplyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Business OS API
 * OpenAPI spec version: 0.1.0
 */

export interface SuggestReplyBody {
  clientMessage: string;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  context?: string | null;
}