# lib/api-zod/src/generated/types/updateClientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Business OS API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateClientBodyStatus } from "./updateClientBodyStatus";

export interface UpdateClientBody {
  name?: string;
  email?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  status?: UpdateClientBodyStatus;
}