# lib/api-zod/src/generated/types/updateDealBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Business OS API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateDealBodyStatus } from "./updateDealBodyStatus";

export interface UpdateDealBody {
  title?: string;
  value?: number;
  status?: UpdateDealBodyStatus;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  notes?: string | null;
}